chore: restructuring directory and updating them snappers :crocodile:
