+++ packages/directory-listing/__tests__/index-spec.js
@@ -2,8 +2,7 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 import toJson from "enzyme-to-json";
-import { Entry } from "../src";
-import Listing from "../src";
+import { Entry, Listing } from "../src";
 
 describe("Listing", () => {
   it("accepts props and renders entries in directory in a table", () => {
@@ -11,7 +10,7 @@ describe("Listing", () => {
       <Listing>
         <Entry>
           <Entry.Icon fileType={"directory"} />
-          <Entry.Name link={"linky"} />
+          <Entry.Name>{"linky"}</Entry.Name>
           <Entry.LastSaved
             last_modified={
               "Fri Jun 22 2018 00:15:55 GMT-0400 (Eastern Daylight Time)"
@@ -29,7 +28,7 @@ describe("Entry", () => {
     const component = shallow(
       <Entry>
         <Entry.Icon fileType={"directory"} />
-        <Entry.Name link={"linky jr"} />
+        <Entry.Name>{"linky jr"}</Entry.Name>
         <Entry.LastSaved
           last_modified={
             "Fri Jun 22 2018 00:15:55 GMT-0400 (Eastern Daylight Time)"
