+++ packages/directory-listing/src/components/entry.js
